requirements.txt
PyVCF==0.6.7
wsgiref==0.1.2
pysam==0.7.8
